+++ apps/api/views/company_views.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
@@ -60,17 +61,17 @@ def create_company_view(request):
         company = create_company_by_company_name_service(user, request.data['name'])
         serializer = CompanySerializer(company, many=False)
         message = 'Запрос выполнен успешно'
-        logger.info(f'{__name__} - {message} / user id:{user.id}')
+        logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -112,21 +113,21 @@ def read_company_view(request, pk):
         if company:
             serializer = CompanySerializer(company, many=False)
             message = 'Запрос выполнен успешно'
-            logger.info(f'{__name__} - {message} / user id:{user.id}')
+            logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
             message = 'Это не ваша company'
-            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
         message = f'Такой company не найдено {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message} / user id:{user.id}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -183,26 +184,26 @@ def update_company_view(request, pk):
             if serializer.is_valid():
                 serializer.save()
                 message = 'Запрос выполнен успешно'
-                logger.info(f'{__name__} - {message} / user id:{user.id}')
+                logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
             message = 'Это не ваша company'
-            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist:
         message = 'Такой company не найдено'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message} / user id:{user.id}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -244,19 +245,19 @@ def delete_company_view(request, pk):
         if company:
             company.delete()
             message = 'Запрос выполнен успешно'
-            logger.info(f'{__name__} - {message} / user id:{user.id}')
+            logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_200_OK)
         else:
             message = 'Это не ваша company'
-            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
         message = f'Такой company не найдено {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message} / user id:{user.id}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)