+++ apps/api/views/qrcode_views.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
@@ -75,31 +76,31 @@ def create_qrcode_view(request):
         if qr_code:
             serializer = QRCodeSerializer(qr_code, many=False, context={"request": request})
             message = 'Запрос выполнен успешно'
-            logger.info(f'{__name__} - {message} / user id:{user.id}')
+            logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             message = 'Это не ваш branch'
-            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
         message = f'Такой branch не найден {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except KeyError as e:
         message = f'Ошибка при обработке запроса. Отсутствует поле {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
     except IntegrityError as e:
         message = f'QRCode уже создан для этого branch {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_406_NOT_ACCEPTABLE)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message} / user id:{user.id}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -140,21 +141,21 @@ def read_qrcode_view(request, pk):
         if qr_code:
             serializer = QRCodeSerializer(qr_code, many=False, context={"request": request})
             message = 'Запрос выполнен успешно'
-            logger.info(f'{__name__} - {message} / user id:{user.id}')
+            logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data=serializer.data, status=status.HTTP_200_OK)
         else:
             message = 'Это не ваш branch'
-            logger.warning(f'{__name__} - {message} / user id:{user.id}')
+            logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
             return Response(data={'detail': message}, status=status.HTTP_403_FORBIDDEN)
 
     except ObjectDoesNotExist as e:
         message = f'Такой QRCode не найден {e}'
-        logger.warning(f'{__name__} - {message} / user id:{user.id}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message} / user id:{user.id}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message} / user id:{user.id}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -186,15 +187,15 @@ def read_all_qrcodes_view(request):
         qr_code_list = get_all_qrcode_service()
         serializer = AllQRCodesSerializer(qr_code_list, many=True, context={"request": request})
         message = 'Запрос выполнен успешно'
-        logger.info(f'{__name__} - {message}')
+        logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(data=serializer.data, status=status.HTTP_200_OK)
 
     except ObjectDoesNotExist as e:
         message = 'QRCodes не найдены'
-        logger.warning(f'{__name__} - {message}')
+        logger.warning(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(data={'detail': message}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
         message = f'Ошибка при обработке запроса {e}'
-        logger.critical(f'{__name__} - {message}')
+        logger.critical(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)