+++ CHANGELOG.md
@@ -2,15 +2,33 @@
 
 ## [Unreleased]
 
-## [0.0.24] - 2022-11-08
+## [0.0.25] - 2022-11-09
 
 ### Добавлено
 
--
+- 
 
 ### Изменено
 
 -
+
+## [0.0.24] - 2022-11-08
+
+### Добавлено
+
+- Yandex tasks [@fringer2423](https://github.com/fringer2423)
+- Zoon tasks
+- 2GIS tasks
+- New app Telebot
+- New model Telebot settings
+- Log to telegram
+- Telegram admin handler
+- Telegram send message task
+
+### Изменено
+
+- Edited all loggers msg [@fringer2423](https://github.com/fringer2423)
+- Task views
 
 ## [0.0.23] - 2022-11-07